pyproject.toml
[project]
name = "project-debater"
version = "0.1.0"
description = "Cambridge Union Exhibition Debate Simulator — multi-agent persona-driven debate simulation"
requires-python = ">=3.11"
dependencies = [
    # --- LLM & Orchestration ---
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",

    # --- Vector Store & Embeddings ---
    "chromadb>=0.5.0",

    # --- Document Processing ---
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",                # Fast PDF text extraction
    "youtube-transcript-api>=0.6.0", # For Cambridge Union YouTube transcripts

    # --- Structured Output ---
    "pydantic>=2.0",

    # --- Observability ---
    "langsmith>=0.1.0",

    # --- Utilities ---
    "python-dotenv>=1.0.0",
    "rich>=13.0",                  # Pretty console output for debate transcripts
    "tiktoken>=0.7.0",            # Token counting

    # --- Cluster Analysis & Visualization ---
    "umap-learn>=0.5.0",          # Dimensionality reduction
    "hdbscan>=0.8.0",            # Density-based clustering
    "plotly>=5.0.0",              # Interactive charts (used in docs/ site)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "ipython>=8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

